move helpers out of endpoints file
